+++ src/algebra/prime/sieve.cpp
@@ -3,26 +3,26 @@
  * @brief Implementation of prime number algorithms defined in sieve.hpp
  */
 
-#include "../include/algebra/prime/sieve.hpp"
+#include "../../../include/algebra/prime/sieve.hpp"
 #include <cmath>
 
 namespace algebra::prime {
 
-    std::vector<int> sieve(int n) {
-        std::vector<int> primes;
+    std::vector<long long unsigned int> sieve(long long unsigned int n) {
+        std::vector<long long unsigned int> primes;
         if (n < 2) return primes;
 
-        std::vector<bool> isPrime(n + 1, true);
+        std::vector<bool> isPrime(static_cast<size_t>(n + 1), true);
         isPrime[0] = isPrime[1] = false;
 
-        for (int i = 2; i * i <= n; ++i) {
+        for (long long unsigned int i = 2; i * i <= n; ++i) {
             if (isPrime[i]) {
-                for (int j = i * i; j <= n; j += i)
+                for (long long unsigned int j = i * i; j <= n; j += i)
                     isPrime[j] = false;
             }
         }
 
-        for (int i = 2; i <= n; ++i)
+        for (long long unsigned int i = 2; i <= n; ++i)
             if (isPrime[i]) primes.push_back(i);
 
         return primes;
